elfloader: use helper variable dtb_size as size_t

+++ elfloader-tool/include/elfloader_common.h
@@ -75,7 +75,7 @@ void load_images(
     unsigned int *num_images,
     void *bootloader_dtb,
     void **chosen_dtb,
-    uint32_t *chosen_dtb_size);
+    size_t *chosen_dtb_size);
 
 /* Platform functions */
 void platform_init(void);
